backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.2.3

# MinIO/S3
minio==7.2.0
boto3==1.29.0

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Security
python-dotenv==1.0.0
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2

# Logging
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.7.1
httpx==0.25.2

# BPJS Integration
pyfhis==0.2.0

# FHIR for SATUSEHAT
fhir.resources==7.0.0
